First working version
Uses docker compose to orchestrate a React frontend that talks via web
sockets to a Python backend, which in-turn talks to the main Python
server process via ZeroMQ. All the network stuff is routed with Nginx.

The backend server uses the dustkid event stream to watch for incoming
replays, and the team hitbox APIs to lookup usernames.