Always reconnect websocket
It turns out that Firefox will close all existing websockets whenever a
link is clicked, even if said link is a custom URI scheme such as
dustforce:// that will not navigate away from the current page.

As a workaround, I will simply try to reconnect the websocket whenever
the connection is closed.